+++ src/main/java/org/realityforge/gwt/symbolmap/SymbolEntryIndexDiff.java
@@ -33,13 +33,11 @@ public final class SymbolEntryIndexDiff
     final SymbolEntryIndexDiff diff = new SymbolEntryIndexDiff( before, after );
     for ( final SymbolEntry entry : before.getSymbolEntries() )
     {
-      final String jsniIdent = entry.getJsniIdent();
-      final boolean isJsniValid = "".equals( jsniIdent );
-      if ( !isJsniValid && null == after.findByJsniIdentifier( jsniIdent ) )
+      if ( entry.isMember() && null == after.findByJsniIdentifier( entry.getJsniIdent() ) )
       {
         diff._missing.add( entry );
       }
-      else if ( isJsniValid && null == after.findByJsName( entry.getJsName() ) )
+      else if ( entry.isType() && null == after.findByJsName( entry.getJsName() ) )
       {
         diff._missing.add( entry );
       }
@@ -50,13 +48,11 @@ public final class SymbolEntryIndexDiff
     }
     for ( final SymbolEntry entry : after.getSymbolEntries() )
     {
-      final String jsniIdent = entry.getJsniIdent();
-      final boolean isJsniValid = "".equals( jsniIdent );
-      if ( !isJsniValid && null == before.findByJsniIdentifier( jsniIdent ) )
+      if ( entry.isMember() && null == before.findByJsniIdentifier( entry.getJsniIdent() ) )
       {
         diff._additional.add( entry );
       }
-      else if ( isJsniValid && null == before.findByJsName( entry.getJsName() ) )
+      else if ( entry.isType() && null == before.findByJsName( entry.getJsName() ) )
       {
         diff._additional.add( entry );
       }